Alterado tipo da coluna "documento", para receber mais caracteres.

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -19,7 +19,7 @@ class CreateUsersTable extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->integer('documento')->unique();
+            $table->bigInteger('documento')->unique();
             $table->string('tipo_conta');
             $table->rememberToken();
             $table->timestamps();
